requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
xgboost
joblib
torchvision
torch
matplotlib
openpyxl
segmentation-models-pytorch
Pillow
streamlit
opencv-python-headless
prince